Move useAnalytics to hooks
