+++ src/components/NewAppReleaseModal.tsx
@@ -8,11 +8,12 @@ import closeIcon from '@assets/images/icn-close.svg';
 import updateImportantIcon from '@assets/images/icn-important-update.svg';
 import updateIcon from '@assets/images/icn-update.svg';
 import { VERSION as currentVersion, GITHUB_RELEASE_NOTES_URL, OS } from '@config';
+import { useAnalytics } from '@hooks';
 import { ANALYTICS_CATEGORIES, GithubService } from '@services/ApiService';
 import { BREAK_POINTS, COLORS } from '@theme';
 import translate from '@translations';
 import { TURL } from '@types';
-import { getFeaturedOS, openLink, useAnalytics } from '@utils';
+import { getFeaturedOS, openLink } from '@utils';
 
 import Modal from './Modal';
