+++ src/components/WalletUnlock/Web3Provider.tsx
@@ -4,12 +4,13 @@ import { Box, Heading, InlineMessage, NewTabLink } from '@components';
 import { Body } from '@components/NewTypography';
 import { IWalletConfig, WALLETS_CONFIG } from '@config';
 import { FormDataActionType as ActionType } from '@features/AddAccount/types';
+import { useAnalytics } from '@hooks';
 import { ANALYTICS_CATEGORIES } from '@services';
 import { NetworkUtils, useNetworks, useSettings } from '@services/Store';
 import { WalletFactory, Web3Wallet } from '@services/WalletService';
 import translate, { translateRaw } from '@translations';
 import { FormData, Network, WalletId } from '@types';
-import { hasWeb3Provider, useAnalytics, useScreenSize } from '@utils';
+import { hasWeb3Provider, useScreenSize } from '@utils';
 import { getWeb3Config } from '@utils/web3';
 
 import './Web3Provider.scss';
