+++ src/components/WalletUnlock/Web3ProviderInstall.tsx
@@ -6,9 +6,10 @@ import AppStoreBadgeIMG from '@assets/images/mobile/app-store-badge.png';
 import GooglePlayBadgeIMG from '@assets/images/mobile/google-play-badge.png';
 import { Box, Heading, NewTabLink } from '@components';
 import { DOWNLOAD_MYCRYPTO_LINK, WALLETS_CONFIG } from '@config';
+import { useAnalytics } from '@hooks';
 import { ANALYTICS_CATEGORIES } from '@services';
 import translate from '@translations';
-import { useAnalytics, useScreenSize } from '@utils';
+import { useScreenSize } from '@utils';
 import './Web3ProviderInstall.scss';
 
 function InstallTrunk() {
