+++ src/features/Dashboard/components/WalletBreakdown/NoAssets.tsx
@@ -4,11 +4,11 @@ import styled from 'styled-components';
 
 import addIcon from '@assets/images/icn-add-assets.svg';
 import { BUY_MYCRYPTO_WEBSITE } from '@config';
+import { useAnalytics } from '@hooks';
 import { ANALYTICS_CATEGORIES } from '@services';
 import { COLORS } from '@theme';
 import translate from '@translations';
 import { openLink } from '@utils';
-import useAnalytics from '@utils/useAnalytics';
 
 const { BLUE_BRIGHT } = COLORS;
