+++ src/features/DownloadApp/DownloadApp.tsx
@@ -8,9 +8,10 @@ import styled from 'styled-components';
 import desktopAppIcon from '@assets/images/icn-desktop-app.svg';
 import { ExtendedContentPanel } from '@components';
 import { DOWNLOAD_MYCRYPTO_LINK, GITHUB_RELEASE_NOTES_URL, OS } from '@config';
+import { useAnalytics } from '@hooks';
 import { ANALYTICS_CATEGORIES, GithubService } from '@services/ApiService';
 import translate from '@translations';
-import { getFeaturedOS, openLink, useAnalytics } from '@utils';
+import { getFeaturedOS, openLink } from '@utils';
 
 import { AppDownloadItem } from './types';
