+++ src/features/Layout/Footer/components/DonateAndSubscribe.tsx
@@ -6,9 +6,9 @@ import CopyToClipboard from 'react-copy-to-clipboard';
 import bitcoin from '@assets/images/bitcoin.png';
 import ether from '@assets/images/ether.png';
 import { donationAddressMap } from '@config';
+import { useAnalytics } from '@hooks';
 import { ANALYTICS_CATEGORIES } from '@services';
 import translate from '@translations';
-import { useAnalytics } from '@utils';
 
 import Subscribe from './Subscribe';
 import './DonateAndSubscribe.scss';
