+++ src/features/Layout/Footer/components/SocialsAndLegal.tsx
@@ -2,10 +2,10 @@ import React, { Component } from 'react';
 
 import { DisclaimerModal, NewTabLink } from '@components';
 import { socialMediaLinks, VERSION } from '@config';
+import { useAnalytics } from '@hooks';
 import { ANALYTICS_CATEGORIES } from '@services';
 import { translateRaw } from '@translations';
 import './SocialsAndLegal.scss';
-import { useAnalytics } from '@utils';
 
 const SocialMediaLink = ({ link, text, icon }: { link: string; text: string; icon?: string }) => {
   const trackSocial = useAnalytics({
