+++ src/features/Layout/Header/Header.tsx
@@ -9,10 +9,11 @@ import LocalIcon from '@components/Icon';
 import { getKBHelpArticle, KB_HELP_ARTICLE, LATEST_NEWS_URL, ROUTE_PATHS } from '@config';
 import { SelectLanguage } from '@features/Drawer/screens';
 import { ScreenLockContext } from '@features/ScreenLock/ScreenLockProvider';
+import { useAnalytics } from '@hooks';
 import { ANALYTICS_CATEGORIES, useFeatureFlags, useSettings } from '@services';
 import { BREAK_POINTS, COLORS, MIN_CONTENT_PADDING } from '@theme';
 import translate, { languages } from '@translations';
-import { openLink, useAnalytics } from '@utils';
+import { openLink } from '@utils';
 
 import { getLinks } from './constants';
