+++ src/features/ScreenLock/ScreenLockForgotPassword.tsx
@@ -7,9 +7,9 @@ import styled from 'styled-components';
 import mainImage from '@assets/images/icn-forgot-password.svg';
 import { ExtendedContentPanel } from '@components';
 import { ROUTE_PATHS } from '@config';
+import { useAnalytics } from '@hooks';
 import { ANALYTICS_CATEGORIES } from '@services';
 import translate, { translateRaw } from '@translations';
-import { useAnalytics } from '@utils';
 
 import { ScreenLockContext } from './ScreenLockProvider';
