+++ src/features/ScreenLock/ScreenLockLocking.tsx
@@ -3,10 +3,11 @@ import React from 'react';
 import { Button } from '@mycrypto/ui';
 import styled from 'styled-components';
 
+import { useAnalytics } from '@hooks';
 import { ANALYTICS_CATEGORIES } from '@services/ApiService';
 import { useSettings } from '@services/Store';
 import translate, { translateRaw } from '@translations';
-import { formatTimeDuration, useAnalytics } from '@utils';
+import { formatTimeDuration } from '@utils';
 
 const MainWrapper = styled.div`
   display: flex;
