+++ src/features/ScreenLock/ScreenLockNew.tsx
@@ -6,9 +6,9 @@ import styled from 'styled-components';
 
 import mainImage from '@assets/images/icn-create-pw.svg';
 import { ExtendedContentPanel, InputField } from '@components';
+import { useAnalytics } from '@hooks';
 import { ANALYTICS_CATEGORIES } from '@services';
 import translate, { translateRaw } from '@translations';
-import { useAnalytics } from '@utils';
 
 import { ScreenLockContext } from './ScreenLockProvider';
