+++ src/hooks/index.ts
@@ -1,1 +1,2 @@
 export { useTxMulti, TxParcel } from './useTxMulti';
+export { default as useAnalytics } from './useAnalytics';
